[Chore] Extract the FFmpeg error mapping from the FFI module
`media/ffi.rs` opens with six helpers that have nothing to do with
transcoding: turning a negative return code into a message, converting
Rust strings and paths into C strings, and validating a rational. They
sit between the caller and every FFI call in the file, so they are the
first thing to move and the least able to break anything -- no ownership,
no lifetimes, one `unsafe` block for `av_strerror`.

They move to `ffi::error`, a leaf, with the four tests that cover them.
